+++ build.gradle.kts
@@ -1,28 +1,46 @@
 plugins {
-    id("jacoco")
-    id("spring-starter.code-metrics")
-    id("spring-starter.java-conventions")
-    id("spring-starter.publishing-conventions")
-    alias(libs.plugins.spotless) apply true
-    alias(libs.plugins.kotlin.jvm) apply false
-    alias(libs.plugins.kotlin.spring) apply false
-    alias(libs.plugins.springboot.dependency.management) apply false
-    alias(libs.plugins.springboot) apply false
-    alias(libs.plugins.openapi.generator) apply false
-    alias(libs.plugins.jooq) apply false
+    alias(libs.plugins.springboot.dependency.management)
+    alias(libs.plugins.springboot)
+    alias(libs.plugins.kotlin.jvm)
+    alias(libs.plugins.kotlin.spring)
+    alias(libs.plugins.spotless)
 }
 
-subprojects {
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        maven { url = uri("https://packages.confluent.io/maven/") }
-        maven {
-            url = uri("https://maven.pkg.github.com/yonatankarp/spring-starter")
-            credentials {
-                username = findProperty("gpr.user")?.toString() ?: System.getenv("GITHUB_ACTOR")
-                password = findProperty("gpr.key")?.toString() ?: System.getenv("GITHUB_TOKEN")
-            }
-        }
+kotlin {
+    jvmToolchain {
+        languageVersion.set(JavaLanguageVersion.of(libs.versions.jvmTarget.get()))
+        target { JavaLanguageVersion.of(libs.versions.jvmTarget.get()) }
     }
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation(libs.bundles.springboot.all)
+    implementation(libs.bundles.kotlin.all)
+}
+
+tasks {
+    jar {
+        enabled = false
+    }
+
+    build {
+        finalizedBy(spotlessApply)
+    }
+
+    test {
+        useJUnitPlatform()
+        finalizedBy(spotlessApply)
+    }
+}
+
+val tasksDependencies = mapOf(
+    "spotlessKotlin" to listOf("compileKotlin", "compileTestKotlin", "test")
+)
+
+tasksDependencies.forEach { (taskName, dependencies) ->
+    tasks.findByName(taskName)?.dependsOn(dependencies)
 }
