--- buildSrc/build.gradle
@@ -1,3 +1,0 @@
-plugins {
-    id("groovy-gradle-plugin")
-}
