--- buildSrc/src/main/groovy/spring-starter.code-metrics.gradle
@@ -1,21 +1,0 @@
-plugins {
-    id("com.diffplug.spotless")
-}
-
-repositories { mavenCentral() }
-
-spotless {
-    kotlin {
-        target(
-                fileTree(projectDir) {
-                    include("**/*.kt")
-                    exclude(
-                            "**/.gradle/**",
-                            "**/build/generated/**"
-                    )
-                }
-        )
-        // see https://github.com/shyiko/ktlint#standard-rules
-        ktlint()
-    }
-}
