--- docker-compose.yml
@@ -1,20 +1,0 @@
-version: "3.9"
-services:
-  postgres:
-    container_name: spring-starter-postgres
-    networks:
-      - proxynet
-    image: postgres:14
-    restart: always
-    environment:
-      POSTGRES_USER: "postgres"
-      POSTGRES_PASSWORD: "secret"
-      POSTGRES_DB: "spring-starter"
-    ports:
-      - "5432:5432"
-    expose:
-      - "5432"
-
-networks:
-  proxynet:
-    name: spring-starter-network
