--- docs/c4/README.md
@@ -1,15 +1,0 @@
-# Architecture diagrams
-
-Architecture is visualized on multiple levels according to a [C4 Model](https://c4model.com/). Diagrams are created as Documentation-as-a-Code. They are written in Structurizr DSL and automatically generated and uploaded into documentation. This way all changes to the architecture can be easily traced back.
-
-## System context diagram (C1)
-
-![structurizr-System-SystemContext.png](structurizr-System-SystemContext.png)
-
-## Container diagram (C2)
-
-![structurizr-System-Container.png](structurizr-System-Container.png)
-
-## Component diagram (C3)
-
-![structurizr-System-Service-Component.png](structurizr-System-Service-Component.png)
