--- spring-starter/src/main/kotlin/com/yonatankarp/spring/starter/Application.kt
@@ -1,11 +1,0 @@
-package com.yonatankarp.spring.starter
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class Application
-
-fun main(args: Array<String>) {
-    runApplication<Application>(*args)
-}
