--- spring-starter/src/main/kotlin/com/yonatankarp/spring/starter/application/ports/HelloWorldPort.kt
@@ -1,5 +1,0 @@
-package com.yonatankarp.spring.starter.application.ports
-
-fun interface HelloWorldPort {
-    suspend fun greet(): String
-}
