--- spring-starter/src/main/kotlin/com/yonatankarp/spring/starter/domain/valueobject/HelloWorld.kt
@@ -1,5 +1,0 @@
-package com.yonatankarp.spring.starter.domain.valueobject
-
-class HelloWorld {
-    fun sayHello(): String = "Hello, World!"
-}
