--- spring-starter/src/main/resources/api/demo-api.yaml
@@ -1,75 +1,0 @@
-openapi: 3.0.3
-info:
-  title: Demo API
-  description: |
-    A dummy hello world API that should be replaced once the service is setup.
-  contact:
-    name: Yonatan Karp-Rudin
-    url: http://yonatankarp.com
-  version: 1.0.0
-
-servers:
-  - url: http://localhost:8080
-    description: The local development environment
-
-tags:
-  - name: DemoApiV1
-
-paths:
-
-  /hello:
-    get:
-      operationId: hello_world
-      summary: Returns hello world to the user
-      description: |
-        This endpoint returning to the client the string "Hello World".
-      tags:
-        - DemoApiV1
-
-      responses:
-        "200":
-          description: |
-            Successfully returning "Hello World" to the client.
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/DemoResponse"
-
-        "500":
-          description: Something unexpected happened on the server side.
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Error"
-
-components:
-
-  schemas:
-
-    DemoResponse:
-      type: object
-      required:
-        - value
-      properties:
-        value:
-          type: string
-          example: "Hello world"
-
-    Error:
-      type: object
-      required:
-        - code
-        - message
-      properties:
-        code:
-          type: string
-          description: Machine-readable error code
-          example: "Not found"
-        message:
-          type: string
-          description: A human-readable error message
-          example: "An error occurred while performing the call"
-        details:
-          type: string
-          description: Unstructured additional data to give more details about the error
-          example: "The error code"
