--- spring-starter/src/main/resources/application.yml
@@ -1,48 +1,0 @@
-server:
-  port: 8080
-
-management:
-  server:
-    port: ${server.port}
-  endpoints:
-    web:
-      exposure:
-        include:
-          - metrics
-          - metrics-requiredMetricName
-          - prometheus
-          - loggers
-          - health
-          - httptrace
-  endpoint:
-    health:
-      probes.enabled: true
-      show-details: always
-      livenessState.enabled: true
-      readinessState.enabled: true
-    metrics.enabled: true
-    prometheus.enabled: true
-    metrics:
-      export.prometheus.enabled: true
-      distribution:
-        percentiles-histogram.http.server:
-          requests: true
-        sla.http.server.requests: 50ms, 100ms, 200ms, 400ms
-        percentiles.http.server:
-          requests: 0.5, 0.9, 0.95, 0.99, 0.999
-
-spring:
-  application:
-    name: spring-starter
-  datasource:
-    url: "jdbc:postgresql://${DB_HOST:localhost}:${DB_PORT:5432}/${DB_SCHEMA:spring-starter}"
-    username: ${DB_USERNAME:postgres}
-    password: ${DB_PASSWORD:secret}
-    driver-class-name: org.postgresql.Driver
-  flyway:
-    enabled: true
-
-springdoc:
-  swagger-ui:
-    path: /api-docs
-    operations-sorter: method
