--- spring-starter/src/main/resources/db/migration/V1.0.0__init_db.sql
@@ -1,6 +1,0 @@
-CREATE TABLE hello_world (
-    id SERIAL PRIMARY KEY,
-    message TEXT NOT NULL
-);
-
-INSERT INTO hello_world (message) VALUES ('Hello World!');
