+++ src/main/kotlin/com/yonatankarp/gracefulerrorhandling/adapters/input/http/rest/HelloWorldHttpAdapter.kt
@@ -1,0 +1,9 @@
+package com.yonatankarp.gracefulerrorhandling.adapters.input.http.rest
+
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+class HelloWorldHttpAdapter {
+    suspend fun helloWorld(): ResponseEntity<Any> = ResponseEntity.ok("Hello World")
+}
