+++ src/main/kotlin/com/yonatankarp/gracefulerrorhandling/application/ports/HelloWorldPort.kt
@@ -1,0 +1,5 @@
+package com.yonatankarp.gracefulerrorhandling.application.ports
+
+fun interface HelloWorldPort {
+    suspend fun greet(): String
+}
