+++ src/main/kotlin/com/yonatankarp/gracefulerrorhandling/application/usecases/HelloWorldUseCase.kt
@@ -1,0 +1,10 @@
+package com.yonatankarp.gracefulerrorhandling.application.usecases
+
+import com.yonatankarp.gracefulerrorhandling.application.ports.HelloWorldPort
+import com.yonatankarp.gracefulerrorhandling.domain.valueobject.HelloWorld
+import org.springframework.stereotype.Service
+
+@Service
+class HelloWorldUseCase : HelloWorldPort {
+    override suspend fun greet() = HelloWorld().sayHello()
+}
