+++ src/main/kotlin/com/yonatankarp/gracefulerrorhandling/domain/valueobject/HelloWorld.kt
@@ -1,0 +1,5 @@
+package com.yonatankarp.gracefulerrorhandling.domain.valueobject
+
+class HelloWorld {
+    fun sayHello(): String = "Hello, World!"
+}
